Change path for saving visualization images
The images generated as a part of the data visualization were previously being saved in the "wwwroot/images/moneylaundering" directory. This commit changes the paths to save these images in the "test" directory instead. This change applies to all plots: suspicious transactions, normal transactions, pie chart overview, and general plot.