Update Home/Index view model to IEnumerable
The commit modifies the Home/Index view to use IEnumerable for the LandingPageViewModel. This allows us to generate a list of landing page summaries, each representing a different country. Furthermore, each summary now includes additional details such as the number of accounts and customers, as well as the total balance per country.