Update image source paths in Inspect.cshtml
Image source paths were hardcoded into the Inspect.cshtml view. This might cause problems with the flexibility of replacing images or updating their names. Hence, the paths are updated to use a dynamic value managed by the PlotNames variable. This improves maintainability of the code. The 'bigplot.png' image is also commented out.