Update data type in DataVisualizationService
The change in the code modifies the data types used for the x and y axis in the DataVisualizationService. This was to accommodate labels instead of DateTime for the x-axis and ensure that the y-axis only contained double values. This is vital in ensuring the accuracy of the graphical representation of our data.