Add return link and customer list to Transfer view
A "Return" link has been added to the "Transfer" view, allowing users to navigate back to the "Details" of the "Cashier" controller easily. Additionally, a table listing customers - if available - has been introduced. This table includes a radio button for selecting the "ToAccountId," along with columns for "Account Id," "Customer Id," "Customer Name," and "Customer Balance."