Refactor Transfer view in BabunBank
The adjustments on the Transfer view include the removal of unnecessary elements like buttons, and table headers for a cleaner and simplified user interface. Further, the presentation of data in cells has been improved for clarity, visually grouping related elements. Account balance is now displayed in a currency format for better readability.