Refactor Filter function in AccountController
The Filter function has been transformed into the GetListOfCustomers method. This change was made to enhance readability and better represent the function's purpose as it retrieves a list of customers. The parameters have been adjusted; functionality checking for pageNumber and pageSize 0 values has been removed and queries related to CreateTransferModel have been updated for clarity.