Refactor AdController and enhance update operation
AdController's constructor and update method were refactored. The constructor no longer requires IConfiguration and the update method now utilizes AutoMapper to simplify the updating of Ad properties. Additionally, a type restriction has been added to the id in the route of the HttpPut action method and the DbContext used to update modified items directly.