Adjust seeding behavior and ad property generation
The constant SeedingVariable has been increased from 100 to 200 to provide a larger dataset. The seed condition changed to add more ads when current count is less than SeedingVariable. The length of the generated description and paragraph range of the content for ads has also been increased bringing variety and more realistic data into the application.