Update assembly informational versions and source link references
This commit updated the assembly informational versions in several files, reflecting the change in the git revision hash. The source link references used for debugging have also been updated to the new revision hash. Additionally, a new data source was added in the `dataSources.xml` which will connect to a database on azure.