Add graph scaling functionality to CashierController
A new method "ScaleGraphs" has been added to the CashierController which allows modifying the scale of graphs in the inspection view. Moreover, a condition has been implemented to check whether the graphs need to be redrawn during account inspection. This provides enhanced visualization options especially useful for larger amounts of data.