Add email confirmation and update static web assets
A new confirmation email field has been added to the customer signup model and the email validation annotation has been applied to it. Additional changes to static web assets have been made, such as the addition of the 'office-test.jpg' and version updates in various AssemblyInfo.cs and .json files.