Add Http client and remove unnecessary code
The changes in this commit include addition of Http client and removal of some redundant lines of code. Specifically, a new singleton service NewsHttpClientEndPoints was configured and added, and an HttpClient was also added with given options. Some comments and unused route mapping code were stripped out for cleaner code.