Updated BabunBank home page layout and content
Improved the home page of BabunBank Web app by enhancing the landing page layout and adding dynamic data. The landing page now presents the number of accounts and customers, as well as the total balance under management dynamically from the system's database. The changes include modification of the layout structure and the content of information presented.