Update routing and method name in AccountController
The route attribute in the AccountController has been modified from "api/[controller]" to just "api". The HTTP GET method has also been renamed from Get to Me and the route for this method is updated to "/me/{id:int}". This change makes the function more specific to get data for the authenticated user.