Update documentation and assembly version numbers
Updated documentation for code files and assembly version numbers. The changes refine the parameters in the Babun-API.xml file and modify AssemblyInformationalVersionAttribute and the .sourcelink.json file in several assemblies. The modification in the AssemblyVersionAttribute promotes correct version management and traceability. A change in the handling of additional files is also reflected in the BabunBank.GeneratedMSBuildEditorConfig.editorconfig file.