Removed Account Details view
The Account Details view ('Details.cshtml' under 'Account' in 'Views') was removed as part of code cleanup or perhaps the view became obsolete. This may be due to changes in the application's functionality or user interface. Confirm if there are corresponding updates in the controller or other dependent code to handle this deletion.