Update cashier details page and add toastr notifications
The modifications in the Details.cshtml file in the Cashier view now handles situations when transaction operation is null. A 'N/A' display value is set when the transaction operation field is null. Furthermore, toastr notifications have been added for better user experience with options set for their appearance, behavior and custom classes.