Change setter properties to init in Transaction.cs
This commit changes most of the properties in the Transaction.cs file from being set to being initialized. This change improves data immutability and ensures that these properties can only be set during initialization, enhancing the security and integrity of the data. The 'Balance' property is the only one that remains with a set accessor since it is expected to change with each transaction.