Implement report date check and three-day sliding average inspection
A new function to check the date of the last report has been added to the MenuService. Also, another method, InspectThreeDaySlidingAverage, has been implemented to inspect accounts based on a three-day sliding window average. In addition, the methods TextMoneyLaundering and PlotAllAccounts have been modified to use the date from the last report.