Refactor CreateAccount method in CashierController
The CreateAccount method within the CashierController has been enhanced. It now includes a process that will confirm account creation with the user and provides meaningful error messaging with a notification when an account cannot be created due to maximum account limit. Additionally, user is redirected to an error page when there is a null return.