Update user deactivation confirmation text in Delete.cshtml
The git diff shows changes made to the layout and text in Delete.cshtml page. The span tag that contained user deactivation confirmation messages is replaced with a div tag, and a padding-top property was added to the second message. Additionally, a sentence was added to the first message about freezing assets.