Add toast notifications to Transfer page
The Transfer view page has been updated with the addition of toast-style notifications. A new script was added to initialize and configure toastr options, and to display a warning message, if it exists in the ViewBag. Notably, these notifications have a custom style and are set to the bottom-center of the screen.