Add exception handling in data seeding operation
An exception handling mechanism was introduced into the data seeding operation in the main program file of the Babun Bank application. This ensures that any failure in the seeding process doesn't halt the rest of the program execution. A comment has been added to a line that may be causing the error.