Refactor pagination and add dropdown in AccountController
Removed the manual computation of totalPageCount and refactored it to be based on Database results. A new dropdown for page size selection has been added, which is populated using values from the CustomerDropdownMenu enum. This will allow users to choose how many results they want to see on a page.