Update assembly info and sourcelink json in multiple projects
This commit includes the update of assembly info and sourcelink json in Babun-API, DataAccessLibrary, and DetectMoneyLaundering projects. It also includes changes in the layout of the BabunBank project. The changes mainly involve adjustment of the runtime version, the reflection assembly informational version attribute, and the URLs for the source repositories.